requirements.txt
alembic==1.15.1
asgiref==3.8.1
Authlib==1.4.1
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-mgmt-core==1.5.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.2.0
azure-mgmt-subscription==3.1.1
babel==2.17.0
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
boto3==1.36.26
botocore==1.36.26
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.43.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dj-database-url==2.3.0
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==24.3
django-sendgrid-v5==1.2.4
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
flask-babel==4.0.0
Flask-Compress==1.17
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.5.2
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.56.0
geopandas==1.0.1
google-api-core==2.24.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonformatter==0.3.4
keyring==25.6.0
kiwisolver==1.4.8
ldap3==2.9.1
libgravatar==1.0.4
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.6.0
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
osmnx==2.0.2
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
passlib==1.7.4
pgadmin4==9.1
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.30.1
psutil==6.1.1
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
PyNaCl==1.5.0
pyogrio==0.10.0
pyotp==2.9.0
pyparsing==3.2.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-http-client==3.3.7
python-socketio==5.12.1
pytz==2025.1
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rich==13.9.4
rsa==4.9
s3transfer==0.11.3
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.3
setuptools==75.9.1
shapely==2.0.7
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
speaklater3==1.4
SQLAlchemy==2.0.39
sqlparse==0.5.3
sshtunnel==0.4.0
starkbank-ecdsa==2.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
Werkzeug==3.1.3
whitenoise==6.9.0
wsproto==1.2.0
WTForms==3.2.1
zstandard==0.23.0
